added spatial search with solr

+++ fitness/server/Fitness.Personalization/Services/IEventDataService.cs
@@ -6,6 +6,6 @@ namespace Sitecore.HabitatHome.Fitness.Personalization.Services
 {
     public interface IEventDataService
     {
-        IEnumerable<Item> GetAll(Database database, string[] profileNames, int take, int skip, float latitude, float longitude, out int totalSearchResults);
+        IEnumerable<Item> GetAll(Database database, string[] profileNames, int take, int skip, double latitude, double longitude, out int totalSearchResults);
     }
 }
